Update test framework and VCR
Update the test framework to use `pytest`.
Update the VCR library to use `vcrpy`.
Update contribution file and `tox.ini` files accordingly.

Using `pytest` allows the use of the library `pytest-vcr`
which records any interrraction of test simply by
adding the annotation `@pytest.mark.vcr` on the test case.

Used fixtures to setup/clean test cases.
Each class uses a fixture to setup its client and create
the necessary spreadsheet. This fixture is called once per class.

The `WorksheetTest` class uses a second fixture to clear the spreadsheet
after each test.

Each test runs in a different instance of the class, we can't benefit
from the use of instance attribute, we must use class attribute to share
resources (like the gspread client and the spreadsheet and the sheet).

Fixtures can't be anotated with the `@pytest.mark.vcr` to record their
interrations, it must be done explicitly.

Updated the github action workflow to run the test suite and fail if
the test suite fails.