Copy comments when during spreadsheet copy
When copying a spreadsheet, if `copy_comments` is True
copy all the comments from original spreadsheet.

When requesting all the comments, the API limit the maximum size of
a page to 100 comments.
Request only the necessary fields from the list of current comments,
to create the new comments iterate over all the comments and push them directly,
no need to change anything.

Recorded the new requests for the tests suite.

closes #978