add method list protected ranges
Add a new method to list all protected ranges.
This method can't be tested in our test suite because it requires to
create a new spreadsheet with a protected range. This requires us to add
the email address used to record the test interactions to create a valid
protected range. This leads to my personal service account leaked in GH.
not cool :-(

It has been tested manually and works fine.

Closes #999