added AppAssertionCredentials to valid oauth2
It appears that Google Colab has changed how it supplies default credentials using the GoogleCredentials.get_application_default(). Now when you authorize through colab--google.colab.auth--the class name (GoogleCredentials.get_application_default().__class__.__name__) is titled AppAssertionCredentials. This is not a valid type according to gspread. I added this in the valid names for the class.