Add method to transfer / accept ownership of a spreadsheet
The process to transfer ownership of a file has been updated.
It now requires an acceptation from the new owner.

Add a method to transfer the ownership, which sets the right parameters
on the current permission.

add a method to accept the ownership, which set the new type as owner
on the current permission.

The API currently does not accept creating new permissions with the
parameter `pendingOwner` set to `true`. One must create first the permission
then edit it.

closes #1053