Fix function `numericise_all`
Function `numericise_all` can ignore some columns
and skip the value at that specific index.

previously used to
1. get the values to ignore on that row
2. iterate over each value of the row
3. if the value is in the list of ignoring values skip it.

Issue: if a row contains twice the same value and the user would like
to ignore one of the two then the other one is ignored too.

Instead use the index to compare if the value at that index should
be skipped or not.

Closes #1069