Prevent crash when passing `None` to `numericise_all`
When a user passes `None` as a list to ignore to the function
`numericise_all` it crashes because it tries to iterate over the list.

cover that case be setting the ignored list to an empty list if not set.

Later solution will be to add typing to this project.

closes #1118