Retrieve modifiedTime and createdTime when missing
When opening a file using it's ID we don't have the modifiedTime and createdTime.
When requested get it using a single API call using the file ID,
in this case if the file is not in the user's drive we can still read it.
Previously using the file listing API returns only the files in the users's drive

Or sometimes we open shared files or public files.

Using the file ID allows us to get the metadata regardless of where the file is stored.

closes #1185