Remplace named tuples with enums
Most of the parameters for the API are `ENUM`s

Replace the named tuples that are not fitted for that
with enums.

Each enum hold one of the choices.
The actual string value must be extracted in order
to pass strings only to the API using the function `extract_enum_value`
using the `enum.value` field if the enum is not `None`.

this forces us to filter all arguments that are passed when they
can possibly be `None`.

thouhgt about setting the default value for each instead
but this changes the API call made to the API and that could lead
to unwanted behavior for the users.

closes #1222