Fix missing method in HTTP Client
Add missing methods to the HTTP client.

update code so the gspread Client and the spreadsheet use
these new methods in order to:
- insert
- list
- remove

a permission.

We can't test these function automatically as this requires
exposing personal information in the recorded requests.

We'll add more tests later when we find a way.

closes #1385

Signed-off-by: Alexandre Lavigne <lavigne958@gmail.com>